requirements.txt
asgiref==3.2.10
boto3==1.15.11
Django==3.1.1
django-admin-ordering==0.12.0
django-environ==0.4.5
django-storages==1.10.1
gunicorn==20.0.4
honcho==1.0.1
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
python-slugify==4.0.1
sqlparse==0.3.1
